+++ tetris_gymnasium/wrappers/action.py
@@ -14,13 +14,13 @@ from tetris_gymnasium.components.tetromino import Tetromino
 from tetris_gymnasium.envs import Tetris
 
 
-class GroupedActionsVector(gym.ObservationWrapper):
+class GroupedActions(gym.ObservationWrapper):
     """Observation wrapper that groups the actions into placements.
 
     The action space is the width of the board times 4 (4 rotations).
     """
 
-    def __init__(self, env: Tetris):
+    def __init__(self, env: Tetris, observation_wrappers: "list[gym.ObservationWrapper]" = None):
         """Initializes the wrapper.
 
         Args:
@@ -28,93 +28,26 @@ class GroupedActionsVector(gym.ObservationWrapper):
         """
         super().__init__(env)
         self.action_space = Discrete((env.unwrapped.width) * 4)
+
+        grouped_env_shape = (env.unwrapped.width * 4,)
+        single_env_shape = observation_wrappers[-1].observation_space.shape if observation_wrappers else env.observation_space["board"].shape
+
         self.observation_space = Box(
             low=0,
             high=env.unwrapped.height * env.unwrapped.width,
-            shape=(
-                4
-                * (
-                    # env.unwrapped.width * env.unwrapped.width + # height
-                    env.unwrapped.width  # max_height
-                    #                         env.unwrapped.width # holes
-                ),
-            ),
+            shape=(grouped_env_shape + single_env_shape),
             dtype=np.float32,
         )
 
         self.legal_actions_mask = np.ones(self.action_space.n)
+        self.observation_wrappers = observation_wrappers
 
-    # def calculate_height(self, board):
-    #     """Calculate the height of each column in the board."""
-    #     board = copy.deepcopy(board)
-    #     board = board[0:-self.env.unwrapped.padding,self.env.unwrapped.padding:-self.env.unwrapped.padding]
-    #     height = np.zeros(board.shape[1], dtype=int)
-    #     for col in range(board.shape[1]):
-    #         for row in range(board.shape[0]):
-    #             if board[row, col] != 0:
-    #                 height[col] = board.shape[0] - row
-    #                 break
-    #     return height
 
-    def calculate_height(self, board):
-        """Calculate the height of each column in the board."""
-        # Slicing the board to remove padding
-        board = board[
-            0 : -self.env.unwrapped.padding,
-            self.env.unwrapped.padding : -self.env.unwrapped.padding,
-        ]
+    def xr_to_action(self, x, r):
+        return x * 4 + r
 
-        # Create a mask where board is not equal to 0
-        mask = board != 0
-
-        # Get the indices of the first non-zero element in each column
-        height = np.argmax(mask, axis=0)
-
-        # Columns with no blocks should have height 0
-        height[np.all(mask == 0, axis=0)] = 0
-
-        # For columns with blocks, calculate the height from the bottom of the board
-        valid_heights = np.any(mask, axis=0)
-        height[valid_heights] = board.shape[0] - height[valid_heights]
-
-        return height
-
-    def calculate_max_height(self, height):
-        """Calculate the maximum height among all columns."""
-        return np.max(height)
-
-    def calculate_holes(self, board):
-        """Calculate the number of holes in the board."""
-        board = copy.deepcopy(board)
-        board = board[
-            0 : -self.env.unwrapped.padding,
-            self.env.unwrapped.padding : -self.env.unwrapped.padding,
-        ]
-
-        if np.all(board == np.ones_like(board)):
-            return board.shape[0] * board.shape[1]
-
-        num_holes = 0
-        for col in zip(*board):
-            row = 0
-            while row < len(board) and col[row] == 0:
-                row += 1
-            num_holes += len([x for x in col[row + 1 :] if x == 0])
-        return num_holes
-
-    def get_bumpiness_and_height(self, board):
-        board = np.array(board)
-        mask = board != 0
-        invert_heights = np.where(
-            mask.any(axis=0), np.argmax(mask, axis=0), self.env.unwrapped.height
-        )
-        heights = self.env.unwrapped.height - invert_heights
-        total_height = np.sum(heights)
-        currs = heights[:-1]
-        nexts = heights[1:]
-        diffs = np.abs(currs - nexts)
-        total_bumpiness = np.sum(diffs)
-        return total_bumpiness, total_height
+    def action_to_xr(self, action):
+        return action // 4, action % 4
 
     def collision_with_frame(self, tetromino: Tetromino, x: int, y: int) -> bool:
         """Check if the tetromino collides with the board at the given position.
@@ -135,7 +68,6 @@ class GroupedActionsVector(gym.ObservationWrapper):
         board_subsection = self.env.unwrapped.board[slices]
 
         # Check collision using numpy element-wise operations.
-        result = np.any(board_subsection[tetromino.matrix > 0] == 1)
         return np.any(board_subsection[tetromino.matrix > 0] == 1)
 
     def observation(self, observation):
@@ -171,10 +103,9 @@ class GroupedActionsVector(gym.ObservationWrapper):
                     y += 1
 
                 # append to results
-                xx = (x - self.env.unwrapped.padding) * 4 + (r + 1) % 4
                 if self.collision_with_frame(t, x, y):
                     self.legal_actions_mask[
-                        (x - self.env.unwrapped.padding) * 4 + (r + 1) % 4
+                        self.xr_to_action(x - self.env.unwrapped.padding, (r + 1) % 4)
                     ] = 0
                     grouped_board_obs.append(np.ones_like(board_obs))
                 elif not self.env.unwrapped.collision(t, x, y):
@@ -185,27 +116,25 @@ class GroupedActionsVector(gym.ObservationWrapper):
                     # regular game over
                     grouped_board_obs.append(np.ones_like(board_obs))
 
-        # concat the results
+        # Apply wrappers
+        if self.observation_wrappers is not None:
+            for i, observation in enumerate(grouped_board_obs):
+                # Recreate the original environment observation
+                grouped_board_obs[i] = {
+                    "board": observation,
+                    "holder": holder_obs,
+                    "queue": queue_obs,
+                }
+
+                # Validate that observations are equal
+                assert grouped_board_obs[i].keys() == self.env.unwrapped.observation_space.keys()
+
+                # Apply wrappers to all the original observations
+                for wrapper in self.observation_wrappers:
+                    grouped_board_obs[i] = wrapper.observation(grouped_board_obs[i])
+
         grouped_board_obs = np.array(grouped_board_obs)
-
-        height_obs = np.array(
-            [self.calculate_height(board) for board in grouped_board_obs]
-        )
-        max_height_obs = np.array(
-            [self.calculate_max_height(height) for height in height_obs]
-        )
-        # holes_obs = np.array([self.calculate_holes(board) for board in grouped_board_obs])
-        # # hb_obs = np.array([self.get_bumpiness_and_height(board) for board in grouped_board_obs])
-        #
-        # # flatten all observations to one array
-        # height_obs = height_obs.flatten()
-        # max_height_obs = max_height_obs.flatten()
-        # holes_obs = holes_obs.flatten()
-        # # hb_obs = hb_obs.flatten()
-
-        # obs = np.concatenate((height_obs, max_height_obs, holes_obs)).astype(np.float32)
-        obs = max_height_obs.astype(np.float32)
-        return obs
+        return grouped_board_obs
 
     def step(self, action):
         """Performs the action.
@@ -216,22 +145,21 @@ class GroupedActionsVector(gym.ObservationWrapper):
         Returns:
             The observation, reward, game over, truncated, and info.
         """
-        x = action // 4
-        r = action % 4
+        x, r = self.action_to_xr(action)
 
         if self.legal_actions_mask[action] == 0:
-            # observations = np.ones(self.observation_space.shape) * self.observation_space.high
-            # reward = self.env.unwrapped.rewards.invalid_action
-            # game_over = True
-            # truncated = False
-            # info = {"action_mask": self.legal_actions_mask}
-            # return observations, reward, game_over, truncated, info
-            obs, reward, game_over, truncated, info = self.env.unwrapped.step(
-                self.env.unwrapped.actions.no_op
-            )
+            observations = np.ones(self.observation_space.shape) * self.observation_space.high
             reward = self.env.unwrapped.rewards.invalid_action
-            info["action_mask"] = self.legal_actions_mask
-            return self.observation(obs), reward, game_over, truncated, info
+            game_over = True
+            truncated = False
+            info = {"action_mask": self.legal_actions_mask}
+            return observations, reward, game_over, truncated, info
+            # obs, reward, game_over, truncated, info = self.env.unwrapped.step(
+            #     self.env.unwrapped.actions.no_op
+            # )
+            # reward = self.env.unwrapped.rewards.invalid_action
+            # info["action_mask"] = self.legal_actions_mask
+            # return self.observation(obs), reward, game_over, truncated, info
 
         new_tetromino = copy.deepcopy(self.env.unwrapped.active_tetromino)
 
@@ -244,13 +172,14 @@ class GroupedActionsVector(gym.ObservationWrapper):
         # Apply rotation and movement (x,y)
         self.env.unwrapped.x = x
         self.env.unwrapped.active_tetromino = new_tetromino
-        observation, reward, game_over, truncated, info = self.env.step(
+
+        # Perform the action
+        observation, reward, game_over, truncated, info = self.env.unwrapped.step(
             self.env.unwrapped.actions.hard_drop
         )
 
-        obs = self.observation(observation)
         info["action_mask"] = self.legal_actions_mask
-        return obs, reward, game_over, truncated, info
+        return self.observation(observation), reward, game_over, truncated, info
 
     def reset(
         self, *, seed: "int | None" = None, options: "dict[str, Any] | None" = None
@@ -268,72 +197,3 @@ class GroupedActionsVector(gym.ObservationWrapper):
         observation, info = self.env.reset(seed=seed, options=options)
         info["action_mask"] = self.legal_actions_mask
         return self.observation(observation), info
-
-    def get_rgb(self, observation):
-        """Observation wrapper that displays all observations (board, holder, queue) as one single RGB Image.
-
-        The observation contains the board on the left, the queue on the top right and the holder on the bottom right.
-        """
-        # Board
-        board_obs = observation["board"]
-        # Holder
-        holder_obs = observation["holder"]
-        # Queue
-        queue_obs = observation["queue"]
-
-        max_size = holder_obs.shape[0]
-        max_len = max(holder_obs.shape[1], queue_obs.shape[1])
-
-        # make holder and queue same length by adding optional padding
-        holder_obs = np.hstack(
-            (holder_obs, np.ones((max_size, max_len - holder_obs.shape[1])))
-        )
-        queue_obs = np.hstack(
-            (queue_obs, np.ones((max_size, max_len - queue_obs.shape[1])))
-        )
-
-        # add vertical padding between the board and the holder/queue
-        v_padding = np.ones((board_obs.shape[0] - 2 * max_size, max_len))
-        cnn_extra = np.vstack((queue_obs, v_padding, holder_obs))
-
-        stack = np.hstack((board_obs, cnn_extra)).astype(np.integer)
-
-        # Convert to RGB
-        rgb = np.zeros((stack.shape[0], stack.shape[1], 3))
-        colors = np.array(
-            list(p.color_rgb for p in self.env.unwrapped.pixels), dtype=np.uint8
-        )
-        rgb[...] = colors[stack]
-
-        return rgb.astype(np.uint8)
-
-    def render(self) -> "RenderFrame | list[RenderFrame] | None":
-        """Renders the environment in various formats.
-
-        This render function is different from the default as it uses the values from :func:`observation`  to render
-        the environment.
-        """
-        matrix = self.get_rgb(self.env.unwrapped._get_obs())
-
-        if self.render_mode == "human" or self.render_mode == "rgb_array":
-            if self.render_mode == "rgb_array":
-                return matrix
-
-            if self.render_mode == "human":
-                if self.env.unwrapped.window_name is None:
-                    self.env.unwrapped.window_name = "Tetris Gymnasium"
-                    cv2.namedWindow(
-                        self.env.unwrapped.window_name, cv2.WINDOW_GUI_NORMAL
-                    )
-                    assert self.observation_space.shape is not None
-                    h, w = (
-                        self.observation_space.shape[0],
-                        self.observation_space.shape[1],
-                    )
-                    cv2.resizeWindow(self.env.unwrapped.window_name, w * 10, h * 10)
-                cv2.imshow(
-                    self.env.unwrapped.window_name,
-                    cv2.cvtColor(matrix, cv2.COLOR_RGB2BGR),
-                )
-
-        return None
