+++ tetris_gymnasium/wrappers/observation.py
@@ -1,4 +1,5 @@
 """Observation wrapper module for the Tetris Gymnasium environment."""
+import copy
 from typing import Any
 
 import cv2
@@ -101,11 +102,10 @@ class RgbObservation(gym.ObservationWrapper):
 
         return None
 
+class FeatureVectorObservation(gym.ObservationWrapper):
+    """Observation wrapper that returns the feature vector as the observation.
 
-class GroupedActionRgbObservation(gym.ObservationWrapper):
-    """Observation wrapper that displays all observations (board, holder, queue) as one single RGB Image.
-
-    The observation contains the board on the left, the queue on the top right and the holder on the bottom right.
+    The feature vector contains the board, the queue and the holder.
     """
 
     def __init__(self, env: Tetris):
@@ -115,111 +115,84 @@ class GroupedActionRgbObservation(gym.ObservationWrapper):
             low=0,
             high=len(env.unwrapped.tetrominoes),
             shape=(
-                env.unwrapped.height_padded * 4,
-                (
-                        env.unwrapped.width_padded
-                        + max(env.unwrapped.holder.size, env.unwrapped.queue.size)
-                        * env.unwrapped.padding
-                )
-                * env.unwrapped.width,
-                3,
+                (env.unwrapped.width + 1,)
             ),
             dtype=np.uint8,
         )
 
-        self.obs = None
+    def calculate_height(self, board):
+        """Calculate the height of each column in the board."""
+        # Slicing the board to remove padding
+        board = board[
+                0: -self.env.unwrapped.padding,
+                self.env.unwrapped.padding: -self.env.unwrapped.padding,
+                ]
+
+        # Create a mask where board is not equal to 0
+        mask = board != 0
+
+        # Get the indices of the first non-zero element in each column
+        height = np.argmax(mask, axis=0)
+
+        # Columns with no blocks should have height 0
+        height[np.all(mask == 0, axis=0)] = 0
+
+        # For columns with blocks, calculate the height from the bottom of the board
+        valid_heights = np.any(mask, axis=0)
+        height[valid_heights] = board.shape[0] - height[valid_heights]
+
+        return height
+
+    def calculate_max_height(self, height):
+        """Calculate the maximum height among all columns."""
+        return np.max(height)
+
+    def calculate_holes(self, board):
+        """Calculate the number of holes in the board."""
+        board = copy.deepcopy(board)
+        board = board[
+                0: -self.env.unwrapped.padding,
+                self.env.unwrapped.padding: -self.env.unwrapped.padding,
+                ]
+
+        if np.all(board == np.ones_like(board)):
+            return board.shape[0] * board.shape[1]
+
+        num_holes = 0
+        for col in zip(*board):
+            row = 0
+            while row < len(board) and col[row] == 0:
+                row += 1
+            num_holes += len([x for x in col[row + 1:] if x == 0])
+        return num_holes
+
+    def get_bumpiness_and_height(self, board):
+        board = np.array(board)
+        mask = board != 0
+        invert_heights = np.where(
+            mask.any(axis=0), np.argmax(mask, axis=0), self.env.unwrapped.height
+        )
+        heights = self.env.unwrapped.height - invert_heights
+        total_height = np.sum(heights)
+        currs = heights[:-1]
+        nexts = heights[1:]
+        diffs = np.abs(currs - nexts)
+        total_bumpiness = np.sum(diffs)
+        return total_bumpiness, total_height
 
     def observation(self, observation):
-        """Observation wrapper that displays all observations (board, holder, queue) as one single RGB Image.
+        """Observation wrapper that returns the feature vector as the observation.
 
-        The observation contains the board on the left, the queue on the top right and the holder on the bottom right.
+        The feature vector contains the board, the queue and the holder.
         """
         # Board
         board_obs = observation["board"]
-        # Holder
-        holder_obs = observation["holder"]
-        # Queue
-        queue_obs = observation["queue"]
 
-        max_size = holder_obs.shape[0]
-        max_len = max(holder_obs.shape[1], queue_obs.shape[1])
+        height = self.calculate_height(board_obs)
+        max_height = self.calculate_max_height(height)
 
-        # make holder and queue same length by adding optional padding
-        holder_obs = np.hstack(
-            (holder_obs, np.ones((max_size, max_len - holder_obs.shape[1])))
-        )
-        queue_obs = np.hstack(
-            (queue_obs, np.ones((max_size, max_len - queue_obs.shape[1])))
-        )
+        # holes_obs = np.array([self.calculate_holes(board) for board in grouped_board_obs])
+        # hb_obs = np.array([self.get_bumpiness_and_height(board) for board in grouped_board_obs])
 
-        v_padding = np.ones((board_obs.shape[1] - 2 * max_size, max_len))
-        cnn_extra = np.vstack((queue_obs, v_padding, holder_obs))
-
-        # add vertical padding between the board and the holder/queue
-        boards = np.array(
-            [np.hstack((o, cnn_extra)).astype(np.integer) for o in board_obs]
-        )
-
-        # Create 2D representation of all boards
-        n_rotations = 4
-        rotated_groups = np.split(boards, boards.shape[0] // n_rotations)
-        matrix = np.hstack([np.vstack(r) for r in rotated_groups])
-
-        # Convert to rgb
-        colors = np.array(
-            list(p.color_rgb for p in self.env.unwrapped.pixels), dtype=np.uint8
-        )
-        matrix_rgb = colors[matrix]
-        self.obs = matrix_rgb.astype(np.uint8)
-
-        return self.obs
-
-    def reset(
-            self, *, seed: "int | None" = None, options: "dict[str, Any] | None" = None
-    ) -> "tuple[dict[str, Any], dict[str, Any]]":
-        """Resets the environment.
-
-        Args:
-            seed: The seed to use for the random number generator.
-            options: The options to use for the environment.
-
-        Returns:
-            The observation and info.
-        """
-        self.obs = None
-        observation, info = self.env.reset(seed=seed, options=options)
-        self.obs = self.observation(observation)
-        return self.observation(observation), info
-
-    def render(self) -> "RenderFrame | list[RenderFrame] | None":
-        """Renders the environment in various formats.
-
-        This render function is different from the default as it uses the values from :func:`observation`  to render
-        the environment.
-        """
-        matrix = self.obs
-        if matrix is None:
-            return None
-
-        if self.render_mode == "human" or self.render_mode == "rgb_array":
-            if self.render_mode == "rgb_array":
-                return matrix
-
-            if self.render_mode == "human":
-                if self.env.unwrapped.window_name is None:
-                    self.env.unwrapped.window_name = "Tetris Gymnasium"
-                    cv2.namedWindow(
-                        self.env.unwrapped.window_name, cv2.WINDOW_GUI_NORMAL
-                    )
-                    assert self.observation_space.shape is not None
-                    h, w = (
-                        self.observation_space.shape[0],
-                        self.observation_space.shape[1],
-                    )
-                    cv2.resizeWindow(self.env.unwrapped.window_name, w * 2, h * 2)
-                cv2.imshow(
-                    self.env.unwrapped.window_name,
-                    cv2.cvtColor(matrix, cv2.COLOR_RGB2BGR),
-                )
-
-        return None
+        observation["board"] = np.array([*height, max_height])
+        return observation ["board"]
